♻️ Refactor: Deprecated 명시 Legacy 전환
