+++ feature/onboarding/src/commonMain/kotlin/com/ondot/onboarding/legacy/OnboardingEvent.kt
@@ -1,4 +1,4 @@
-package com.ondot.onboarding
+package com.ondot.onboarding.legacy
 
 import com.ondot.ui.base.Event
